Shipping-Service/requirements.txt
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.11.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.75.2
grpcio==1.56.0
grpcio-tools==1.56.0
h11==0.14.0
idna==3.4
kombu==5.3.1
lxml==4.9.1
outcome==1.2.0
prompt-toolkit==3.0.38
protobuf==4.23.3
pycparser==2.21
pydantic==1.10.9
pymongo==4.4.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.0
selenium==4.1.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.17.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
urllib3-secure-extra==0.1.0
uvicorn==0.17.4
uvloop==0.17.0
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==3.5.4
wsproto==1.2.0
